refactor: Calendar component
